Authorizator and Entity access control. So far for view and list actions.

+++ src/SimpleApi/EntityController.php
@@ -3,6 +3,9 @@
 namespace OndraKoupil\AppTools\SimpleApi;
 
 use Exception;
+use NotORM_Result;
+use OndraKoupil\AppTools\SimpleApi\Auth\EntityAuthorizatorInterface;
+use OndraKoupil\AppTools\SimpleApi\Auth\IdentityExtractorInterface;
 use OndraKoupil\Tools\Arrays;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
@@ -15,11 +18,29 @@ class EntityController {
 	 */
 	protected $manager;
 
+	/**
+	 * @var EntityAuthorizatorInterface|null
+	 */
+	protected $authorizator;
+
+	/**
+	 * @var IdentityExtractorInterface|null
+	 */
+	private $identityExtractor;
+
 	function __construct(
-		EntityManagerInterface $manager
+		EntityManagerInterface $manager,
+		IdentityExtractorInterface $identityExtractor = null,
+		EntityAuthorizatorInterface $authorizator = null
 	) {
 
 		$this->manager = $manager;
+		$this->authorizator = $authorizator;
+		$this->identityExtractor = $identityExtractor;
+
+		if ($this->authorizator and !$this->identityExtractor) {
+			throw new Exception('To use Authorizator in EntityController, an IdentityExtractor is required!');
+		}
 	}
 
 	function respondWithNothing(ResponseInterface $response): ResponseInterface {
@@ -38,15 +59,30 @@ class EntityController {
 		return $r;
 	}
 
+	function respondUnauthorized(ResponseInterface $response): ResponseInterface {
+		return $this->respondWithError($response, 403, 'You are not allowed to do this.');
+	}
+
 	function list(ServerRequestInterface $request, ResponseInterface $response): ResponseInterface {
 		$parts = $this->getPartsFromRequest($request);
-		return $this->respondWithJson($response, $this->manager->getAllItems($parts));
+		$restriction = null;
+		if ($this->authorizator and $this->identityExtractor) {
+			$user = $this->identityExtractor->getUser($request);
+			$restriction = $this->authorizator->createListRestriction($user);
+		}
+		$allItems = $this->manager->getAllItems($parts, $restriction);
+		return $this->respondWithJson($response, $allItems);
 	}
 
 	function view(ServerRequestInterface $request, ResponseInterface $response, string $id): ResponseInterface {
 		try {
 			$parts = $this->getPartsFromRequest($request);
 			$data = $this->manager->getItem($id, $parts);
+			if ($this->authorizator and $this->identityExtractor) {
+				if (!$this->authorizator->canView($this->identityExtractor->getUser($request), $data)) {
+					return $this->respondUnauthorized($response);
+				}
+			}
 			return $this->respondWithJson($response, $data);
 		} catch (ItemNotFoundException $e) {
 			return $this->respondWithError($response, 404, 'Item with ID ' . $e->notFoundId . ' was not found.');
@@ -58,6 +94,13 @@ class EntityController {
 			$ids = array_values(array_map('trim', explode(',', $idAsStringWithCommas)));
 			$parts = $this->getPartsFromRequest($request);
 			$data = $this->manager->getManyItems($ids, $parts);
+
+			if ($this->authorizator and $this->identityExtractor) {
+				if (!$this->authorizator->canViewMany($this->identityExtractor->getUser($request), $data)) {
+					return $this->respondUnauthorized($response);
+				}
+			}
+
 			return $this->respondWithJson($response, $data);
 		} catch (ItemNotFoundException $e) {
 			return $this->respondWithError($response, 404, 'Item with ID ' . $e->notFoundId . ' was not found.');
@@ -70,11 +113,17 @@ class EntityController {
 		if (!$data) {
 			return $this->respondWithError($response, 400, 'Missing body with data.');
 		}
+		if ($this->authorizator) {
+			throw new Exception('Not implemented yet!');
+		}
 		$item = $this->manager->createItem($data);
 		return $this->respondWithJson($response, $item);
 	}
 
 	function delete(ServerRequestInterface $request, ResponseInterface $response, string $id): ResponseInterface {
+		if ($this->authorizator) {
+			throw new Exception('Not implemented yet!');
+		}
 		try {
 			$this->manager->deleteItem($id);
 			return $this->respondWithNothing($response);
@@ -84,6 +133,9 @@ class EntityController {
 	}
 
 	function deleteMany(ServerRequestInterface $request, ResponseInterface $response): ResponseInterface {
+		if ($this->authorizator) {
+			throw new Exception('Not implemented yet!');
+		}
 		$body = $request->getParsedBody();
 		if (!($body['id'] ?? null)) {
 			return $this->respondWithError($response, 400, 'Missing [id] parameter with ID\'s to delete.');
@@ -99,6 +151,9 @@ class EntityController {
 	}
 
 	function edit(ServerRequestInterface $request, ResponseInterface $response, string $id): ResponseInterface {
+		if ($this->authorizator) {
+			throw new Exception('Not implemented yet!');
+		}
 		$data = $request->getParsedBody();
 		if (!$data) {
 			return $this->respondWithError($response, 400, 'Missing body with data.');
@@ -112,6 +167,9 @@ class EntityController {
 	}
 
 	function clone(ServerRequestInterface $request, ResponseInterface $response, string $id): ResponseInterface {
+		if ($this->authorizator) {
+			throw new Exception('Not implemented yet!');
+		}
 		try {
 			$created = $this->manager->cloneItem($id);
 			return $this->respondWithJson($response, $created);
